Move update_yaml_dict from shub.config to shub.utils
