+++ shub/login.py
@@ -4,8 +4,10 @@ import requests
 from six.moves import input
 from six.moves.urllib.parse import urljoin
 
-from shub.config import load_shub_config, ShubConfig, update_yaml_dict
+from shub.config import (load_shub_config, GLOBAL_SCRAPINGHUB_YML_PATH,
+                         ShubConfig)
 from shub.exceptions import AlreadyLoggedInException
+from shub.utils import update_yaml_dict
 
 
 HELP = """
@@ -31,7 +33,7 @@ def cli():
         suggestion=conf.apikeys.get('default'),
         endpoint=global_conf.endpoints.get('default'),
     )
-    with update_yaml_dict() as conf:
+    with update_yaml_dict(GLOBAL_SCRAPINGHUB_YML_PATH) as conf:
         conf.setdefault('apikeys', {})
         conf['apikeys']['default'] = key
