+++ tests/test_utils.py
@@ -8,7 +8,10 @@ import os
 import stat
 import sys
 import unittest
+import textwrap
 import time
+
+import yaml
 
 from mock import Mock, MagicMock, patch
 from click.testing import CliRunner
@@ -406,6 +409,30 @@ class UtilsTest(unittest.TestCase):
                                   rsp=rsp, verbose=True)
         self.assertEqual(last_logs, [
             b"line1", b'{"status":"ok","fieldK":"fieldV"}'])
+
+    def test_update_yaml_dict(self):
+        YAML_BEFORE = textwrap.dedent("""\
+            a:
+              unrelated: dict
+            b:
+              key1: val1
+              key2: val2
+        """)
+        DICT_EXPECTED = {
+            'a': {'unrelated': 'dict'},
+            'b': {'key1': 'newval1', 'key2': 'val2', 'key3': 'val3'}
+        }
+        runner = CliRunner()
+        with runner.isolated_filesystem():
+            with open('conf.yml', 'w') as f:
+                f.write(YAML_BEFORE)
+            with utils.update_yaml_dict('conf.yml') as conf:
+                conf['b']['key1'] = 'newval1'
+                conf['b']['key3'] = 'val3'
+            with open('conf.yml', 'r') as f:
+                self.assertEqual(yaml.safe_load(f), DICT_EXPECTED)
+                f.seek(0)
+                self.assertIn("key1: newval1", f.read())
 
 
 if __name__ == '__main__':
